pyproject.toml
[tool.poetry]
name = "ch-exporter"
version = "23.03.01"
description = ""
authors = ["Jannik Hoffjann <hi@hfjn.de>"]
readme = "README.md"
packages = [{include = "ch_exporter"}]

[tool.poetry.dependencies]
python = "^3.11"
prometheus-client = "^0.16.0"
pydantic = "^1.10.4"
aiohttp = "^3.8.3"
aiochclient = {extras = ["aiohttp"], version = "^2.2.0"}
black = "^22.12.0"
pyyaml = "^6.0"
loguru = "^0.6.0"
pendulum = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"